evert: update fstab
* and add firmware class path to cmdline

+++ recovery/root/sbin/prepdecrypt.sh
@@ -90,8 +90,13 @@ cp /v/lib64/hw/android.hardware.boot@1.0-impl.so /vendor/lib64/hw/
 cp /v/lib64/hw/android.hardware.gatekeeper@1.0-impl-qti.so /vendor/lib64/hw/
 cp /v/lib64/hw/android.hardware.keymaster@3.0-impl-qti.so /vendor/lib64/hw/
 
-cp /v/manifest.xml /vendor/
-cp /v/compatibility_matrix.xml /vendor/
+if [ -f /v/manifest.xml ]; then
+  cp /v/manifest.xml /vendor/
+  cp /v/compatibility_matrix.xml /vendor/
+else
+  cp /v/etc/vintf/manifest.xml /vendor/
+  cp /v/etc/vintf/compatibility_matrix.xml /vendor/
+fi
 
 relink /v/bin/hw/android.hardware.boot@1.0-service
 relink /v/bin/hw/android.hardware.gatekeeper@1.0-service-qti
